requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.1
python-multipart==0.0.20
streamlit==1.44.1
plotly==6.0.1
pillow>=10.3,<11
ultralytics==8.3.81
numpy==2.1.1
rasterio==1.4.3
geopandas==1.1.1
pyogrio==0.11.1
shapely==2.1.1
pyproj==3.7.1
requests>=2.31
httpx==0.28.1
requests-toolbelt==1.0.0
